add config.py
