--- webapp/server.py
@@ -1,18 +1,0 @@
-from flask import Flask, render_template
-from webapp.weather import weather_by_city
-from  webapp.python_org_news import get_python_news
-
-
-app = Flask(__name__)
-
-
-@app.route('/')
-def index():
-    title = "Новости Python"
-    weather = weather_by_city('Moscow,Russia')
-    news_list = get_python_news()
-    return render_template('index.html', page_title=title, weather=weather, news_list=news_list)
-
-
-if __name__ == "__main__":
-    app.run()
