+++ webapp/weather.py
@@ -1,9 +1,10 @@
+from flask import current_app
 import requests
 
 def weather_by_city(city_name):
-    weather_url = 'http://api.worldweatheronline.com/premium/v1/weather.ashx'
+    weather_url = current_app.config["WEATHER_URL"]
     params = {
-        "key": "0811b97620ea4a8982c110542222401",
+        "key": current_app.config["WEATHER_API_KEY"],
         "q": city_name,
         "format": "json",
         "num_of_days": 1,
